+++ src/Command/Make/ExitException.php
@@ -20,4 +20,5 @@ class ExitException extends \RuntimeException
     const STATUS_CHANGELOG_TYPE_UNKNOWN = 2;
     const STATUS_NO_CHANGES = 3;
     const STATUS_CHANGELOG_FILE_NOT_FOUND = 4;
+    const STATUS_SCRIPT_CANNOT_RUN = 5;
 }