+++ tests/unit/Chimney/Command/Make/OutputMessageTest.php
@@ -12,6 +12,7 @@
 namespace Plista\Chimney\Test\Unit\Command\Make;
 
 use Plista\Chimney\Command\Make\OutputMessage;
+use Plista\Chimney\Command\Make\PlaceholderManager as P;
 
 /**
  *
@@ -35,6 +36,7 @@ class OutputMessageTest extends \PHPUnit_Framework_TestCase
 
     protected function setUp() {
         $this->message = new OutputMessage();
+        $this->message->appendN("beginning");
         $this->hrSingle = OutputMessage::HR_SINGLE;
         $this->hrDouble = OutputMessage::HR_DOUBLE;
     }
@@ -45,6 +47,7 @@ class OutputMessageTest extends \PHPUnit_Framework_TestCase
     public function appendChangelogInfo() {
         $this->message->appendChangelogInfo('myChangelogAddon', 'myChangelogPath');
         $this->assertEquals(<<<EOT
+beginning
 
 <info>{$this->hrDouble}
 Generated changelog:
@@ -53,6 +56,77 @@ myChangelogAddon
 <comment>{$this->hrSingle}
 The changelog was added to myChangelogPath. You don't need to edit it manually.
 {$this->hrDouble}</comment>
+
+EOT
+            ,
+            $this->message->get()
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function appendHintDebian() {
+        $version = '4.0.0';
+        $package = 'plista-chimney';
+        $changelog = '/usr/share/chimney/debian/changelog';
+
+        $placeholderManager = new P();
+        $placeholderManager
+            ->collect(P::VERSION, $version)
+            ->collect(P::PACKAGE_NAME, $package)
+            ->collect(P::CHANGELOG_FILE, $changelog);
+        $this->message->appendHintDebian($placeholderManager);
+        $this->assertEquals(<<<"EOT"
+beginning
+
+<info>====================
+Release commands:
+====================</info>
+    git checkout next
+    git pull
+    git commit -m "{$package} ($version) $changelog"
+    git push
+    git checkout master
+    git pull
+    git merge next
+    git push
+    git checkout next
+<comment>--------------------
+Copy and paste these command into your console for quicker releasing.
+====================</comment>
+
+EOT
+            ,
+            $this->message->get()
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function appendHintMd() {
+        $version = '4.0.0';
+        $changelog = '/usr/share/chimney/debian/changelog';
+
+        $placeholderManager = new P();
+        $placeholderManager
+            ->collect(P::VERSION, $version)
+            ->collect(P::CHANGELOG_FILE, $changelog);
+        $this->message->appendHintMd($placeholderManager);
+        $this->assertEquals(<<<"EOT"
+beginning
+
+<info>====================
+Release commands:
+====================</info>
+    git commit -m "Update changelog #ign" {$changelog}
+    git tag {$version}
+    git push
+    git push --tags
+<comment>--------------------
+Copy and paste these command into your console for quicker releasing.
+====================</comment>
 
 EOT
             ,